feat(discovery): add transitive dependency graph generation and Mermaid diagram support
