+++ src/DevOpsMigrationPlatform.Infrastructure/Modules/Discovery/MermaidDiagramBuilder.cs
@@ -5,13 +5,13 @@ using System.Text;
 using System.Text.RegularExpressions;
 using DevOpsMigrationPlatform.Abstractions.Models;
 
-namespace DevOpsMigrationPlatform.CLI.Commands.Discovery;
+namespace DevOpsMigrationPlatform.Infrastructure.Modules.Discovery;
 
 /// <summary>
 /// Generates Mermaid flowchart diagrams from project dependency records.
 /// Cross-organisation targets are visually distinguished with orange styling.
 /// </summary>
-internal sealed class MermaidDiagramBuilder
+public sealed class MermaidDiagramBuilder
 {
     private readonly IEnumerable<ProjectDependencyRecord> _pairs;
 
@@ -39,7 +39,7 @@ internal sealed class MermaidDiagramBuilder
             if (pair.LinkScope == LinkScope.CrossOrganisation)
                 allProjects.Add($"{pair.TargetOrganisation}/{pair.TargetProject ?? "remote"}");
             else if (!string.IsNullOrWhiteSpace(pair.TargetProject))
-                allProjects.Add(pair.TargetProject);
+                allProjects.Add(pair.TargetProject!);
         }
 
         foreach (var project in allProjects)
@@ -56,7 +56,7 @@ internal sealed class MermaidDiagramBuilder
                 ? $"{pair.TargetOrganisation}/{pair.TargetProject ?? "remote"}"
                 : pair.TargetProject;
 
-            if (!string.IsNullOrWhiteSpace(targetName) && nodeMap.TryGetValue(targetName, out var targetId))
+            if (!string.IsNullOrWhiteSpace(targetName) && nodeMap.TryGetValue(targetName!, out var targetId))
             {
                 // Edge with label: source -->|"N links"| target
                 sb.AppendLine($"    {sourceId} -->|\"{pair.LinkCount}\"| {targetId}");
